+++ crypto/crypto.go
@@ -1,5 +1,7 @@
 package crypto
 
+import "github.com/aptos-labs/aptos-go-sdk/bcs"
+
 // Signer a generic interface for any kind of signing
 type Signer interface {
 	// ToHex if a private key, it's bytes, if it's not a private key
@@ -7,19 +9,17 @@ type Signer interface {
 	ToHex
 
 	// Sign signs a transaction and returns an associated authenticator
-	Sign(msg []byte) (authenticator Authenticator, err error)
+	Sign(msg []byte) (authenticator *Authenticator, err error)
 
 	// AuthKey gives the AuthenticationKey associated with the signer
 	AuthKey() *AuthenticationKey
 }
 
 // PrivateKey a generic interface for a signing private key
+// This is not serializable, because this doesn't go on-chain
 type PrivateKey interface {
 	Signer
-	ToHex
-	FromHex
-	ToBytes
-	FromBytes
+	CryptoMaterial
 
 	// PubKey Retrieve the public key for signature verification
 	PubKey() PublicKey
@@ -27,6 +27,8 @@ type PrivateKey interface {
 
 // PublicKey a generic interface for a public key associated with the private key
 type PublicKey interface {
+	bcs.Struct
+	CryptoMaterial
 	ToHex
 	FromHex
 	ToBytes
@@ -37,6 +39,13 @@ type PublicKey interface {
 
 	// Verify verifies a message with the public key
 	Verify(msg []byte, sig []byte) bool
+}
+
+type CryptoMaterial interface {
+	ToHex
+	FromHex
+	ToBytes
+	FromBytes
 }
 
 type FromHex interface {
@@ -54,6 +63,6 @@ type FromBytes interface {
 }
 
 type ToBytes interface {
-	// Bytes loads the key from bytes
+	// CryptoMaterial loads the key from bytes
 	Bytes() []byte
 }